Apply suggestions from code review

+++ tedana/utils.py
@@ -919,7 +919,7 @@ def load_mask(ref_img, mask=None, t2smap=None):
         RepLGR.info("A user-defined mask was applied to the data.")
         mask_img = nb.load(mask)
         # Convert mask to binary (assuming non-zero values indicate mask inclusion)
-        mask = mask_img.get_fdata().astype(np.uint8) > 0
+        mask = mask_img.get_fdata() > 0
         mask_img = nb.Nifti1Image(mask, mask_img.affine, mask_img.header)
     elif t2smap and not mask:
         LGR.info("Assuming user-defined T2* map is masked and using it to generate mask")
@@ -936,7 +936,7 @@ def load_mask(ref_img, mask=None, t2smap=None):
         # Load and convert user-defined mask to NIfTI1 (e.g., AFNI format)
         mask_img = io._convert_to_nifti1(nb.load(mask))
         # Convert mask to binary (assuming non-zero values indicate mask inclusion)
-        mask = mask_img.get_fdata().astype(np.uint8) > 0
+        mask = mask_img.get_fdata() > 0
         mask[t2s_loaded == 0] = 0  # reduce mask based on T2* map
         mask_img = nb.Nifti1Image(mask, mask_img.affine, mask_img.header)
         t2s = apply_mask(t2s_img, mask_img)
